requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.8
django-environ==0.4.1
django-phonenumber-field==1.1.0
django-rest-swagger==2.0.5
djangorestframework==3.4.6
djangorestframework-gis==0.10.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
phonenumbers==8.10.15
phonenumberslite==8.10.15
psycopg2==2.7.1
psycopg2-binary==2.8.3
pycountry==1.20
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
yapf==0.28.0
